% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{hfdl_set_key}
\alias{hfdl_set_key}
\title{Set the HF Data Library API key for this session.}
\usage{
hfdl_set_key(api_key)
}
\arguments{
\item{api_key}{Your API key (from https://hfdatalibrary.com/pages/account).}
}
\description{
Set the HF Data Library API key for this session.
}
